requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
django-restframework==0.0.1
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
